Fix UnfoldingAbstractDatatypesFinite to make theorems more accurate
To make proofs about infinite objects, we just make proofs about arbitrarily
long finite portions of them. When we stop unfolding in unfoldrn, we store the
current iteration state in the output, by converting it to the right type with
an arbitrary function passed by the caller: the alternative would be discarding
the state, but then proofs would maybe be less meaningful.
It's not clear whether proofs on such tree portions extend to infinite trees, but it
seems more likely for these tree portions than for ones discarding the state.

Extend this trick from unfoldrn to unfoldTn.