WIP correct untyped deBrujin indexes
I've translated the standard algorithm, only to notice at the end that
one step relies on negative shifting, which I don't support, and which
is hard to support in Agda — you'd need as argument a proof that some
variables don't exist. I think that's a misdesign: instead, I believe
again that substituting a variable should simply remove it from the
context, as in my previous design. I'm going to try this change next.

This commit contains a big commented out experiment for variable shift,
working but a bit messy. I'll remove this next, but I've kept it because
its messiness motivates a new non-obvious but easier design, with ↑ and
↑′.