TypedSelfRepresentation: Adapt TSR interface
It's probably more reasonable to assume that Expr is parameterized over
a type representation rather than a type --- at least, that's more
common.

The original TSR paper by Rendel, Ostermann and Hofer didn't introduce a
type representation, but that's probably the exception rather than the
rule.

This change happens to imply, annoyingly, that the object language is
dependently-typed, or that `Type` is a kind of type-level type
representations (like, I guess, in Brown-Palsberg).