Exploit diff indexing more to simplify proofs
Do this for IlcListDIff as an example; see comment to understand what
makes the proof simpler.

Also play with irrelevance to show that, in `patch`, we do not actually
need to take as argument the target of the change, and it could be
erased. This is however not too satisfactory.